Remove black bars for empty days in overview.
The overview shows black bars for days with no activities in the
interval from the first and last days having activities in the
selected date range. Not only are these bars unsightly, but the
behaviour is inconsistent because there are no bars for empty days
at the start and end of the date range.

Historically there was some sense to the placeholder rows because
they were actionable (for adding activities?) but this functionality
was lost in the transition to gtk3.