Fix delayed updating of overview
The facts-changed signalling has been broken ever since the removal
of DBus and client.py. As a poor substitute, changes to hamster-lite.db
were monitored and this resulted in an awkward short delay before the
overview was updated. Here the facts-changed signal is fired by the
storage DB module and the file monitoring has been removed.